+++ crates/forge/tests/e2e/debugging.rs
@@ -16,6 +16,19 @@ fn debugging_trace_detailed() {
 }
 
 #[test]
+fn debugging_trace_detailed_fork() {
+    let temp = setup_package("debugging_fork");
+
+    let output = test_runner(&temp)
+        .arg("--trace-verbosity")
+        .arg("detailed")
+        .assert()
+        .code(1);
+
+    assert_stdout_contains(output, test_output(detailed_debugging_trace_message_fork));
+}
+
+#[test]
 fn debugging_trace_standard() {
     let temp = setup_package("debugging");
 
@@ -26,6 +39,19 @@ fn debugging_trace_standard() {
         .code(1);
 
     assert_stdout_contains(output, test_output(standard_debugging_trace_message));
+}
+
+#[test]
+fn debugging_trace_standard_fork() {
+    let temp = setup_package("debugging_fork");
+
+    let output = test_runner(&temp)
+        .arg("--trace-verbosity")
+        .arg("standard")
+        .assert()
+        .code(1);
+
+    assert_stdout_contains(output, test_output(standard_debugging_trace_message_fork));
 }
 
 #[test]
@@ -41,6 +67,19 @@ fn debugging_trace_minimal() {
     assert_stdout_contains(output, test_output(minimal_debugging_trace_message));
 }
 
+#[test]
+fn debugging_trace_minimal_fork() {
+    let temp = setup_package("debugging_fork");
+
+    let output = test_runner(&temp)
+        .arg("--trace-verbosity")
+        .arg("minimal")
+        .assert()
+        .code(1);
+
+    assert_stdout_contains(output, test_output(minimal_debugging_trace_message_fork));
+}
+
 fn test_output(trace_message_fn: fn(&str) -> String) -> String {
     formatdoc! {r"
         [..]Compiling[..]
@@ -48,7 +87,7 @@ fn test_output(trace_message_fn: fn(&str) -> String) -> String {
 
         Collected 2 test(s) from trace_info package
         Running 2 test(s) from tests/
-        [FAIL] trace_info_integrationtest::test_trace::test_debugging_trace_fail
+        [FAIL] trace_info_integrationtest::test_trace::test_debugging_trace_failure
         Failure data:
             (0x1, 0x2, 0x3, 0x4, 0x5)
 
@@ -62,9 +101,9 @@ fn test_output(trace_message_fn: fn(&str) -> String) -> String {
         Tests: 1 passed, 1 failed, 0 skipped, 0 ignored, 0 filtered out
 
         Failures:
-            trace_info_integrationtest::test_trace::test_debugging_trace_fail
+            trace_info_integrationtest::test_trace::test_debugging_trace_failure
         ",
-        debugging_trace_fail = trace_message_fn("fail"),
+        debugging_trace_fail = trace_message_fn("failure"),
         debugging_trace_pass = trace_message_fn("success")
     }
 }
@@ -76,7 +115,7 @@ fn detailed_debugging_trace_message(test_name: &str) -> String {
         │  ├─ [contract name] SimpleContract
         │  ├─ [entry point type] External
         │  ├─ [calldata] array![RecursiveCall {{ contract_address: ContractAddress([..]), payload: array![RecursiveCall {{ contract_address: ContractAddress([..]), payload: array![] }}, RecursiveCall {{ contract_address: ContractAddress([..]), payload: array![] }}] }}, RecursiveCall {{ contract_address: ContractAddress([..]), payload: array![] }}]
-        │  ├─ [storage address] [..]
+        │  ├─ [contract address] [..]
         │  ├─ [caller address] [..]
         │  ├─ [call type] Call
         │  ├─ [call result] success: array![RecursiveCall {{ contract_address: ContractAddress([..]), payload: array![RecursiveCall {{ contract_address: ContractAddress([..]), payload: array![] }}, RecursiveCall {{ contract_address: ContractAddress([..]), payload: array![] }}] }}, RecursiveCall {{ contract_address: ContractAddress([..]), payload: array![] }}]
@@ -84,7 +123,7 @@ fn detailed_debugging_trace_message(test_name: &str) -> String {
         │  │  ├─ [contract name] SimpleContract
         │  │  ├─ [entry point type] External
         │  │  ├─ [calldata] array![RecursiveCall {{ contract_address: ContractAddress([..]), payload: array![] }}, RecursiveCall {{ contract_address: ContractAddress([..]), payload: array![] }}]
-        │  │  ├─ [storage address] [..]
+        │  │  ├─ [contract address] [..]
         │  │  ├─ [caller address] [..]
         │  │  ├─ [call type] Call
         │  │  ├─ [call result] success: array![RecursiveCall {{ contract_address: ContractAddress([..]), payload: array![] }}, RecursiveCall {{ contract_address: ContractAddress([..]), payload: array![] }}]
@@ -92,7 +131,7 @@ fn detailed_debugging_trace_message(test_name: &str) -> String {
         │  │  │  ├─ [contract name] SimpleContract
         │  │  │  ├─ [entry point type] External
         │  │  │  ├─ [calldata] array![]
-        │  │  │  ├─ [storage address] [..]
+        │  │  │  ├─ [contract address] [..]
         │  │  │  ├─ [caller address] [..]
         │  │  │  ├─ [call type] Call
         │  │  │  └─ [call result] success: array![]
@@ -100,7 +139,7 @@ fn detailed_debugging_trace_message(test_name: &str) -> String {
         │  │     ├─ [contract name] SimpleContract
         │  │     ├─ [entry point type] External
         │  │     ├─ [calldata] array![]
-        │  │     ├─ [storage address] [..]
+        │  │     ├─ [contract address] [..]
         │  │     ├─ [caller address] [..]
         │  │     ├─ [call type] Call
         │  │     └─ [call result] success: array![]
@@ -108,7 +147,7 @@ fn detailed_debugging_trace_message(test_name: &str) -> String {
         │     ├─ [contract name] SimpleContract
         │     ├─ [entry point type] External
         │     ├─ [calldata] array![]
-        │     ├─ [storage address] [..]
+        │     ├─ [contract address] [..]
         │     ├─ [caller address] [..]
         │     ├─ [call type] Call
         │     └─ [call result] success: array![]
@@ -116,7 +155,61 @@ fn detailed_debugging_trace_message(test_name: &str) -> String {
            ├─ [contract name] SimpleContract
            ├─ [entry point type] External
            ├─ [calldata] array![0x1, 0x2, 0x3, 0x4, 0x5]
-           ├─ [storage address] [..]
+           ├─ [contract address] [..]
+           ├─ [caller address] [..]
+           ├─ [call type] Call
+           └─ [call result] panic: (0x1, 0x2, 0x3, 0x4, 0x5)
+        "}
+}
+
+fn detailed_debugging_trace_message_fork(test_name: &str) -> String {
+    formatdoc! {r"
+        [test name] trace_info_integrationtest::test_trace::test_debugging_trace_{test_name}
+        ├─ [selector] execute_calls
+        │  ├─ [contract name] forked contract
+        │  ├─ [entry point type] External
+        │  ├─ [calldata] array![RecursiveCall {{ contract_address: ContractAddress([..]), payload: array![RecursiveCall {{ contract_address: ContractAddress([..]), payload: array![] }}, RecursiveCall {{ contract_address: ContractAddress([..]), payload: array![] }}] }}, RecursiveCall {{ contract_address: ContractAddress([..]), payload: array![] }}]
+        │  ├─ [contract address] [..]
+        │  ├─ [caller address] [..]
+        │  ├─ [call type] Call
+        │  ├─ [call result] success: array![RecursiveCall {{ contract_address: ContractAddress([..]), payload: array![RecursiveCall {{ contract_address: ContractAddress([..]), payload: array![] }}, RecursiveCall {{ contract_address: ContractAddress([..]), payload: array![] }}] }}, RecursiveCall {{ contract_address: ContractAddress([..]), payload: array![] }}]
+        │  ├─ [selector] execute_calls
+        │  │  ├─ [contract name] forked contract
+        │  │  ├─ [entry point type] External
+        │  │  ├─ [calldata] array![RecursiveCall {{ contract_address: ContractAddress([..]), payload: array![] }}, RecursiveCall {{ contract_address: ContractAddress([..]), payload: array![] }}]
+        │  │  ├─ [contract address] [..]
+        │  │  ├─ [caller address] [..]
+        │  │  ├─ [call type] Call
+        │  │  ├─ [call result] success: array![RecursiveCall {{ contract_address: ContractAddress([..]), payload: array![] }}, RecursiveCall {{ contract_address: ContractAddress([..]), payload: array![] }}]
+        │  │  ├─ [selector] execute_calls
+        │  │  │  ├─ [contract name] forked contract
+        │  │  │  ├─ [entry point type] External
+        │  │  │  ├─ [calldata] array![]
+        │  │  │  ├─ [contract address] [..]
+        │  │  │  ├─ [caller address] [..]
+        │  │  │  ├─ [call type] Call
+        │  │  │  └─ [call result] success: array![]
+        │  │  └─ [selector] execute_calls
+        │  │     ├─ [contract name] forked contract
+        │  │     ├─ [entry point type] External
+        │  │     ├─ [calldata] array![]
+        │  │     ├─ [contract address] [..]
+        │  │     ├─ [caller address] [..]
+        │  │     ├─ [call type] Call
+        │  │     └─ [call result] success: array![]
+        │  └─ [selector] execute_calls
+        │     ├─ [contract name] forked contract
+        │     ├─ [entry point type] External
+        │     ├─ [calldata] array![]
+        │     ├─ [contract address] [..]
+        │     ├─ [caller address] [..]
+        │     ├─ [call type] Call
+        │     └─ [call result] success: array![]
+        └─ [selector] fail
+           ├─ [contract name] forked contract
+           ├─ [entry point type] External
+           ├─ [calldata] array![0x1, 0x2, 0x3, 0x4, 0x5]
+           ├─ [contract address] [..]
            ├─ [caller address] [..]
            ├─ [call type] Call
            └─ [call result] panic: (0x1, 0x2, 0x3, 0x4, 0x5)
@@ -153,6 +246,36 @@ fn standard_debugging_trace_message(test_name: &str) -> String {
         "}
 }
 
+fn standard_debugging_trace_message_fork(test_name: &str) -> String {
+    formatdoc! {r"
+        [test name] trace_info_integrationtest::test_trace::test_debugging_trace_{test_name}
+        ├─ [selector] execute_calls
+        │  ├─ [contract name] forked contract
+        │  ├─ [calldata] array![RecursiveCall {{ contract_address: ContractAddress([..]), payload: array![RecursiveCall {{ contract_address: ContractAddress([..]), payload: array![] }}, RecursiveCall {{ contract_address: ContractAddress([..]), payload: array![] }}] }}, RecursiveCall {{ contract_address: ContractAddress([..]), payload: array![] }}]
+        │  ├─ [call result] success: array![RecursiveCall {{ contract_address: ContractAddress([..]), payload: array![RecursiveCall {{ contract_address: ContractAddress([..]), payload: array![] }}, RecursiveCall {{ contract_address: ContractAddress([..]), payload: array![] }}] }}, RecursiveCall {{ contract_address: ContractAddress([..]), payload: array![] }}]
+        │  ├─ [selector] execute_calls
+        │  │  ├─ [contract name] forked contract
+        │  │  ├─ [calldata] array![RecursiveCall {{ contract_address: ContractAddress([..]), payload: array![] }}, RecursiveCall {{ contract_address: ContractAddress([..]), payload: array![] }}]
+        │  │  ├─ [call result] success: array![RecursiveCall {{ contract_address: ContractAddress([..]), payload: array![] }}, RecursiveCall {{ contract_address: ContractAddress([..]), payload: array![] }}]
+        │  │  ├─ [selector] execute_calls
+        │  │  │  ├─ [contract name] forked contract
+        │  │  │  ├─ [calldata] array![]
+        │  │  │  └─ [call result] success: array![]
+        │  │  └─ [selector] execute_calls
+        │  │     ├─ [contract name] forked contract
+        │  │     ├─ [calldata] array![]
+        │  │     └─ [call result] success: array![]
+        │  └─ [selector] execute_calls
+        │     ├─ [contract name] forked contract
+        │     ├─ [calldata] array![]
+        │     └─ [call result] success: array![]
+        └─ [selector] fail
+           ├─ [contract name] forked contract
+           ├─ [calldata] array![0x1, 0x2, 0x3, 0x4, 0x5]
+           └─ [call result] panic: (0x1, 0x2, 0x3, 0x4, 0x5)
+        "}
+}
+
 fn minimal_debugging_trace_message(test_name: &str) -> String {
     formatdoc! {r"
         [test name] trace_info_integrationtest::test_trace::test_debugging_trace_{test_name}
@@ -168,5 +291,23 @@ fn minimal_debugging_trace_message(test_name: &str) -> String {
         │     └─ [contract name] SimpleContract
         └─ [selector] fail
            └─ [contract name] SimpleContract
+        "}
+}
+
+fn minimal_debugging_trace_message_fork(test_name: &str) -> String {
+    formatdoc! {r"
+        [test name] trace_info_integrationtest::test_trace::test_debugging_trace_{test_name}
+        ├─ [selector] execute_calls
+        │  ├─ [contract name] forked contract
+        │  ├─ [selector] execute_calls
+        │  │  ├─ [contract name] forked contract
+        │  │  ├─ [selector] execute_calls
+        │  │  │  └─ [contract name] forked contract
+        │  │  └─ [selector] execute_calls
+        │  │     └─ [contract name] forked contract
+        │  └─ [selector] execute_calls
+        │     └─ [contract name] forked contract
+        └─ [selector] fail
+           └─ [contract name] forked contract
         "}
 }