+++ crates/sncast/src/response/declare.rs
@@ -1,0 +1,53 @@
+use conversions::{padded_felt::PaddedFelt, serde::serialize::CairoSerialize};
+use indoc::formatdoc;
+use serde::{Deserialize, Serialize};
+
+use crate::helpers::block_explorer::LinkProvider;
+
+use super::{command::CommandResponse, explorer_link::OutputLink};
+
+#[derive(Clone, Serialize, Deserialize, CairoSerialize, Debug, PartialEq)]
+pub struct DeclareTransactionResponse {
+    pub class_hash: PaddedFelt,
+    pub transaction_hash: PaddedFelt,
+}
+
+impl CommandResponse for DeclareTransactionResponse {}
+
+// TODO(#3391): Update text output to be more user friendly
+// impl Message for SncastMessage<DeclareTransactionResponse> {}
+
+#[derive(Clone, Serialize, Deserialize, CairoSerialize, Debug, PartialEq)]
+pub struct AlreadyDeclaredResponse {
+    pub class_hash: PaddedFelt,
+}
+
+impl CommandResponse for AlreadyDeclaredResponse {}
+
+// TODO(#3391): Update text output to be more user friendly
+// impl Message for SncastMessage<AlreadyDeclaredResponse> {}
+
+#[derive(Clone, Serialize, Deserialize, CairoSerialize, Debug, PartialEq)]
+#[serde(tag = "status")]
+pub enum DeclareResponse {
+    AlreadyDeclared(AlreadyDeclaredResponse),
+    #[serde(untagged)]
+    Success(DeclareTransactionResponse),
+}
+
+impl CommandResponse for DeclareResponse {}
+
+impl OutputLink for DeclareTransactionResponse {
+    const TITLE: &'static str = "declaration";
+
+    fn format_links(&self, provider: Box<dyn LinkProvider>) -> String {
+        formatdoc!(
+            "
+            class: {}
+            transaction: {}
+            ",
+            provider.class(self.class_hash),
+            provider.transaction(self.transaction_hash)
+        )
+    }
+}